requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
cycler==0.11.0
daphne==3.0.2
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
flatbuffers==22.11.23
fonttools==4.25.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
imutils==0.5.4
incremental==21.3.0
joblib==1.1.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
multidict==6.0.2
munkres==1.1.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
pip==22.3.1
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pysqlite3==0.4.7
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
pyu2f==0.1.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
service-identity==21.1.0
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.1.1
threadpoolctl==2.2.0
torch==1.13.0
tornado==6.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wheel==0.37.1
wrapt==1.14.1
yarl==1.8.1
zipp==3.11.0
zope.interface==5.4.0
